feat(search): polish global search palette
