+++ Sources/Search/GlobalSearchResultRow.swift
@@ -1,0 +1,40 @@
+import Foundation
+
+struct GlobalSearchResultRow: Identifiable, Equatable {
+    let hit: SearchIndexHit
+    let query: String
+    let index: Int
+
+    var id: String { hit.id }
+
+    var title: String {
+        let trimmed = hit.title.trimmingCharacters(in: .whitespacesAndNewlines)
+        return trimmed.isEmpty
+            ? String(localized: "globalSearch.untitled", defaultValue: "Untitled")
+            : trimmed
+    }
+
+    var location: String {
+        hit.location.trimmingCharacters(in: .whitespacesAndNewlines)
+    }
+
+    var snippet: String {
+        let trimmed = hit.snippet.trimmingCharacters(in: .whitespacesAndNewlines)
+        return trimmed.isEmpty ? title : trimmed
+    }
+
+    var shortcutLabel: String? {
+        index < 9 ? "⌘\(index + 1)" : nil
+    }
+
+    var systemImageName: String {
+        switch hit.kind {
+        case .browser:
+            return "globe"
+        case .markdown:
+            return "doc.richtext"
+        case .title:
+            return "macwindow"
+        }
+    }
+}
